# spec/generated/openapi.yaml
openapi: 3.1.0
info:
  title: Booking Calendar API
  version: 1.0.0
  description: |-
    Booking Calendar API — лёгкий аналог cal.com.

    В системе две роли и нет ни регистрации, ни авторизации:

    - **Владелец календаря** — единственный заранее заданный профиль (`Owner`).
    Он создаёт типы событий (`EventType`) и просматривает единый список
    предстоящих встреч по всем типам событий.
    - **Гость** — анонимный посетитель. Смотрит список видов брони, открывает
    календарь свободных слотов на 14 дней и создаёт бронирование (`Booking`)
    без создания аккаунта.

    Ключевое правило занятости: на одно и то же время нельзя создать две записи,
    даже если это разные типы событий. Занятость считается по всему календарю
    владельца, а не в рамках одного типа события.

    Окно записи: слоты формируются на `Owner.bookingWindowDays` (по умолчанию 14)
    дней, начиная с текущего момента.
...
paths:
...
  /event-types/{eventTypeId}/slots:
    get:
      operationId: Public_listSlots
      summary: Получить свободные слоты на окно записи
      description: |-
        Календарь свободных слотов выбранного типа события.

        Слоты формируются на окно записи `[from; to)`, по умолчанию — на
        `Owner.bookingWindowDays` (14) дней, начиная с текущего момента.
        Внутри окна берутся только интервалы, кратные `Owner.slotStepMinutes`,
        целиком укладывающиеся в `Owner.workingHours` и не пересекающиеся ни с
        одним существующим бронированием календаря — включая брони других типов
        событий.

        Ошибки: `400 SLOT_OUT_OF_WINDOW`, если `from`/`to` выходят за окно записи
        или `to` не больше `from`; `404 EVENT_TYPE_NOT_FOUND`.
      parameters:
        - name: eventTypeId
          in: path
          required: true
          description: Идентификатор типа события.
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Начало запрашиваемого окна. По умолчанию — текущий момент.
          schema:
            type: string
            format: date-time
          explode: false
        - name: to
          in: query
          required: false
          description: |-
            Конец запрашиваемого окна. По умолчанию — `from + bookingWindowDays`
            (14 дней). Не может выходить за границу окна записи.
          schema:
            type: string
            format: date-time
          explode: false
        - name: onlyAvailable
          in: query
          required: false
          description: Возвращать только свободные слоты. При `false` слоты приходят с `isAvailable`.
          schema:
            type: boolean
            default: true
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotsResponse'
        '400':
          description: Ответ с ошибкой и конкретным HTTP-статусом.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Ответ с ошибкой и конкретным HTTP-статусом.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Guest
components:
...
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Машиночитаемый код ошибки.
        message:
          type: string
          description: Человекочитаемое описание проблемы.
        details:
          type: object
          unevaluatedProperties: {}
          description: Дополнительный контекст ошибки (например, имя невалидного поля).
      description: Стандартное тело ошибки.
...
    DaySlots:
      type: object
      required:
        - date
        - slots
      properties:
        date:
          type: string
          format: date
          description: Дата дня в часовом поясе владельца.
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
          description: Слоты этого дня. Пустой массив, если свободного времени нет.
      description: Слоты одного календарного дня в часовом поясе владельца.
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - EVENT_TYPE_NOT_FOUND
        - EVENT_TYPE_ID_CONFLICT
        - BOOKING_NOT_FOUND
        - SLOT_TAKEN
        - SLOT_OUT_OF_WINDOW
        - SLOT_NOT_ALIGNED
      description: Машиночитаемый код ошибки домена.
...
    Slot:
      type: object
      required:
        - start
        - end
        - isAvailable
      properties:
        start:
          type: string
          format: date-time
          description: Начало слота.
        end:
          type: string
          format: date-time
          description: 'Окончание слота: `start + EventType.durationMinutes`.'
        isAvailable:
          type: boolean
          description: |-
            Свободен ли слот. `false`, если слот пересекается с любым существующим
            бронированием календаря — независимо от его типа события.
      description: |-
        Слот — вычисляемый интервал в календаре владельца. Слоты не хранятся:
        они строятся с шагом `Owner.slotStepMinutes` внутри `Owner.workingHours`
        на длину `EventType.durationMinutes`.
    SlotsResponse:
      type: object
      required:
        - eventTypeId
        - durationMinutes
        - timezone
        - windowStart
        - windowEnd
        - days
      properties:
        eventTypeId:
          type: string
          description: Тип события, для которого построены слоты.
        durationMinutes:
          type: integer
          format: int32
          description: Длительность встречи, использованная при построении слотов.
        timezone:
          $ref: '#/components/schemas/Timezone'
          description: Часовой пояс владельца, в котором сгруппированы дни.
        windowStart:
          type: string
          format: date-time
          description: Начало окна записи (по умолчанию — текущий момент).
        windowEnd:
          type: string
          format: date-time
          description: Конец окна записи (по умолчанию — `windowStart + bookingWindowDays`).
        days:
          type: array
          items:
            $ref: '#/components/schemas/DaySlots'
          description: Дни окна записи по возрастанию, включая дни без свободных слотов.
      description: Календарь свободных слотов для выбранного типа события.
    Timezone:
      type: string
      description: Часовой пояс в формате IANA, например `Europe/Moscow`.
      examples:
        - Europe/Moscow
...
servers:
  - url: /api/v1
    description: Базовый путь API
    variables: {}